docker-deploy/web-app/requirements.txt
asgiref==3.3.1
asn1crypto==0.24.0
certifi==2018.1.18
chardet==3.0.4
cryptography==2.1.4
Django==3.1.6
django-crispy-forms==1.11.0
django-filter==2.4.0
django-multiselectfield==0.1.12
httplib2==0.9.2
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
netifaces==0.10.4
psycopg2==2.8.6
pycrypto==2.6.1
pytz==2021.1
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
six==1.11.0
sqlparse==0.4.1
urllib3==1.22